split definitions into damage effects
