+++ xlive/Blam/Engine/objects/damage_effects.cpp
@@ -1,0 +1,2 @@
+#include "stdafx.h"
+#include "damage_effects.h"
